OnlineApproval/requirements.txt
asgiref==3.2.7
asn1crypto==1.3.0
astroid==2.4.1
attrs==19.3.0
bcrypt==3.1.7
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
cryptography==2.9.2
defusedxml==0.6.0
diff-match-patch==20200713
Django==3.0.5
django-import-export==2.3.0
django-widget-tweaks==1.4.8
django-wkhtmltopdf==3.3.0
et-xmlfile==1.0.1
idna==2.10
isort==4.3.21
jdcal==1.4.1
jsonfield==3.1.0
lazy-object-proxy==1.4.3
lockfile==0.12.2
lxml==4.5.2
MarkupPy==1.14
mccabe==0.6.1
odfpy==1.4.1
openpyxl==3.0.4
oscrypto==1.2.0
paramiko==2.7.1
pdfkit==0.6.1
Pillow==7.1.2
psycopg2==2.8.4
py-moneyed==0.8.0
pycparser==2.20
pylint==2.5.2
pylint-django==2.0.15
pylint-plugin-utils==0.6
PyNaCl==1.4.0
pyOpenSSL==19.1.0
pytz==2019.3
PyYAML==5.3.1
requests==2.24.0
six==1.14.0
sqlparse==0.3.1
tablib==2.0.0
toml==0.10.0
typed-ast==1.4.1
urllib3==1.25.9
wrapt==1.12.1
xlrd==1.2.0
xlwt==1.3.0
